Cleanup switch statement and constructor
Apparently old school php used a different type of constructor, and
because my function name is the same as the class name, it was trying to
use my function as a constructor upon initialization.  Furthermore, this
meant an uninitialized value was being passed to my switch statement,
and I didn't have a proper default catch.