Setup router
Couldn't get the annotation-based router to work, so I switched to the
"normal" router.

I didn't realize this, but there was an apache setting I had to setup to
To make it recongize .htaccess files.  I hadn't noticed before because
my index.php was in /var/www instead of /var/www/public.

I updated the setting, and now everything seems to be working.