Setup scaffolding for screen scraper command
I decided to store the data for the complexes in a persistent json file
instead of grabbing them everytime because they are unlike to change
often, and it seems dangerous to depend too much on the html sctructure.

Determining the availability of appartments likely has to be done in
real time.