Finish setting up db, screen-scraper, and models
The database and configuration services are now working.

The screen scraper CLI command is now working; runningn it populates the
database with appartment complexes it finds on the url specified in the
config.  Currently it deletes all previous quieries upon being run!

I filled in part of the appartment service with database queires.